refactor: allow start from project root
