--- packages/core/src/env-set/append-dot-env.ts
@@ -1,7 +1,0 @@
-import { appendFileSync } from 'fs';
-
-const appendDotEnv = (key: string, value: string) => {
-  appendFileSync('.env', `${key}=${value}\n`);
-};
-
-export default appendDotEnv;
