+++ packages/core/src/env-set/parameters.ts
@@ -1,2 +1,3 @@
-const parameters = process.argv.slice(2);
-export const noInquiry = parameters.includes('--no-inquiry');
+const parameters = new Set(process.argv.slice(2));
+export const noInquiry = parameters.has('--no-inquiry');
+export const fromRoot = parameters.has('--from-root');
